Name changes for clarity
The purpose of a state machine definition(HsmDefinition) is to capture all the
structural details of a state machine. It contains the states (including sub
and orthogonal HSMs), events and the state transition table.

A state machine executor (HsmExecutor) implements all the "dynamic" details -
onEntry, onExit and execute methods. In addition, there is a dispatch
method that is responsible for delivering an event to the innermost
active HSM in a hierarchy of HSMs.

Based on the above reasoning, StateMachineDef was renamed to
HsmDefinition and StateMachine was renamed to HsmExecutor. The other
renamed types should be more obvious.